+++ init.h
@@ -4,6 +4,6 @@
 #define MAX_SEED 16384
 
 int RandSeed(int);
-double** InitMyMatrix(int, int, double);
+double** InitMyMatrix(int, int, int);
 
 #endif
